requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
cffi==1.16.0
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.1.3
fastapi==0.104.1
greenlet==3.0.1
h11==0.14.0
idna==3.4
passlib==1.7.4
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
pymacaroons==0.13.0
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
